+++ package.py
@@ -1,6 +1,6 @@
 name = "bootycall"
 
-version = "0.6.4"
+version = "0.7.0"
 
 description = "An interface for managing rez environments and resolving packages."